requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pymodbus>=3.6.0
pysnmp>=6.1.0
aiosqlite>=0.20.0
paho-mqtt>=1.6.1,<3.0
adafruit-circuitpython-bme280>=2.6.0
adafruit-circuitpython-dht>=4.0.0
RPi.GPIO>=0.7.1; platform_machine=="aarch64"
smbus2>=0.4.3
pydantic>=2.0.0
python-multipart>=0.0.9
pyserial>=3.5          # WIO Terminal USB-Serial (SHT40_USB sensor type)
wsproto>=1.0.0         # Explicitly required by uvicorn ws="wsproto" implementation
